reorganize project structure and add READMEs
