+++ subsystem/beam_pulse/beam_pulse.py
@@ -1,0 +1,9 @@
+
+
+class BeamPulseSubsystem:
+    """
+    Controls beam pulse timing and characteristics
+    """
+
+    def __init__(self):
+        # TODO: Implement
