+++ subsystem/deflection_monitor/deflection_monitor.py
@@ -1,0 +1,8 @@
+
+
+class DeflectionMonitorSubsystem:
+    """
+    Monitors beam deflection and position
+    """
+    def __init__(self):
+        # TODO: Implement
